fix(types): extract CacheStore to break Message/Socket cycle
Moving CacheStore types into cache-store.ts removes the circular import
between Message.ts and Socket.ts so SocketConfig.telemetry and
MessageGenerationOptions resolve correctly for consumers.
